requirements-dev.txt
# RPi.GPIO >= 0.7.1
Mock.GPIO
fastapi >= 0.78.0
uvicorn >= 0.18.1
gunicorn >= 20.1.0
PyYAML >= 6.0
python-jose >= 3.3.0
python-multipart >= 0.0.5